restructured module into seperate files, called by main.py
