+++ fcpx_marker_tool/common/helpers.py
