+++ fcpx_marker_tool/parsers/fcp7parser.py
@@ -1,0 +1,5 @@
+class FCP7Parser:
+
+    def __init__(self, xml_root):
+        self.xml_root = xml_root
+        print("FCP7 support will be added in a future version.")
